pyproject.toml
[project]
name = "infrascope"
version = "0.1.0"
description = "IT infrastructure monitoring and administration platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.14",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30",
    "bcrypt>=4.2",
    "fastapi>=0.129.0",
    "httpx>=0.28",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.7",
    "pydantic[email]>=2.10",
    "pyjwt[crypto]>=2.10",
    "pysnmp-lextudio>=6.0",
    "python-multipart>=0.0.18",
    "redis>=5.0",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.34",
    "paramiko>=3.5",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "celery>=5.6.2",
    "kafka-python>=2.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "pytest>=8.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pyright]
pythonVersion = "3.13"
